added DataDisplay Component, and React-Router-DOM
