+++ src/DataDisplay/index.js
@@ -1,0 +1,21 @@
+import React from 'react'
+
+class DataDisplay extends React.Component {
+  constructor(props){
+    super(props)
+    this.state = {
+
+    }
+  }
+
+  render() {
+
+    return (
+      <h1>Data display!</h1>
+    )
+
+
+  }
+}
+
+export default DataDisplay
